requirements.txt
# Smart Shamba — Production Requirements
# Python 3.12.x
# ═══════════════════════════════════════════════════════════════════════

# ── Web Framework (FastAPI) ──────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Database: PostgreSQL + pgvector ──────────────────────────────────
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
asyncpg>=0.30.0
pgvector>=0.3.0
sqlalchemy[asyncio]>=2.0.0

# ── IoT / MQTT ──────────────────────────────────────────────────────
paho-mqtt>=2.1.0

# ── Machine Learning: PyTorch + Vision + Serialisation ───────────────
torch>=2.0.0
torchvision>=0.15.0
safetensors>=0.4.0
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.1.0
imbalanced-learn>=0.11.0

# ── Async Task Queue ────────────────────────────────────────────────
celery[redis]>=5.3.0

# ── AI: Google Gemini + RAG Embeddings ───────────────────────────────
google-genai>=1.0.0
sentence-transformers>=3.0.0

# ── Earth Engine & Geospatial ────────────────────────────────────────
earthengine-api>=0.1.400
rasterio>=1.3.0
geopandas>=0.14.0
folium>=0.17.0
xarray>=2024.1.0

# ── SMS: Africa's Talking ───────────────────────────────────────────
requests>=2.31.0

# ── Utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.0
schedule>=1.2.0
python-dateutil>=2.8.0
bcrypt>=4.1.0
certifi>=2023.7.0
Pillow>=10.0.0
httpx>=0.27.0

# ── Testing ──────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0

# ── Visualization ───────────────────────────────────────────────────
plotly>=5.22.0

# ── Legacy (optional — only needed for migration script) ─────────────
# pymongo>=4.6.0
africastalking>=1.2.0
# openai>=1.0.0
# scikit-learn>=1.3.0
# streamlit>=1.50.0
# streamlit-folium>=0.15.0
# streamlit-lottie>=0.0.5
# flask>=3.0.0
# gunicorn>=21.2.0
# sendgrid>=6.11.0

